Dogfood the mods list
