+++ library/core/crash_info/src/main/java/org/quiltmc/qsl/crash/impl/CrashInfoImpl.java
@@ -1,0 +1,27 @@
+package org.quiltmc.qsl.crash.impl;
+
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.loader.api.FabricLoader;
+import net.fabricmc.loader.api.ModContainer;
+import org.quiltmc.qsl.crash.api.CrashReportEvents;
+
+public final class CrashInfoImpl implements ModInitializer {
+	@Override
+	public void onInitialize() {
+		CrashReportEvents.SYSTEM_DETAILS.register(details -> details.addSection("Quilt Mods", () -> {
+			StringBuilder builder = new StringBuilder();
+
+			for (ModContainer mod : FabricLoader.getInstance().getAllMods()) {
+				var metadata = mod.getMetadata();
+
+				builder.append("\n\t\t%s: %s %s".formatted(
+						metadata.getId(),
+						metadata.getName(),
+						metadata.getVersion().getFriendlyString())
+				);
+			}
+
+			return builder.toString();
+		}));
+	}
+}
