+++ library/core/crash_info/src/main/resources/fabric.mod.json
@@ -14,6 +14,11 @@
   "authors": [
     "QuiltMC"
   ],
+  "entrypoints": {
+    "main": [
+      "org.quiltmc.qsl.crash.impl.CrashInfoImpl"
+    ]
+  },
   "depends": {
     "quilt_loader": ">=0.13.1-rc.10"
   },
